+++ src/main/java/com/arangodb/internal/util/RequestUtils.java
@@ -1,7 +1,7 @@
 /*
  * DISCLAIMER
  *
- * Copyright 2016 ArangoDB GmbH, Cologne, Germany
+ * Copyright 2018 ArangoDB GmbH, Cologne, Germany
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,27 +18,33 @@
  * Copyright holder is ArangoDB GmbH, Cologne, Germany
  */
 
-package com.arangodb.internal.net;
+package com.arangodb.internal.util;
 
-import java.io.IOException;
+import com.arangodb.internal.net.AccessType;
+import com.arangodb.velocystream.Request;
 
 /**
  * @author Mark Vollmary
  *
  */
-public interface HostHandler {
+public final class RequestUtils {
 
-	Host get(HostHandle hostHandle, AccessType accessType);
+	public static final String HEADER_ALLOW_DIRTY_READ = "X-Arango-Allow-Dirty-Read";
 
-	void success();
+	private RequestUtils() {
+		super();
+	}
 
-	void fail();
+	public static AccessType determineAccessType(final Request request) {
+		if (request.getHeaderParam().containsKey(HEADER_ALLOW_DIRTY_READ)) {
+			return AccessType.DIRTY_READ;
+		}
+		switch (request.getRequestType()) {
+		case GET:
+			return AccessType.READ;
+		default:
+			return AccessType.WRITE;
+		}
+	}
 
-	void reset();
-
-	void confirm();
-
-	void close() throws IOException;
-
-	void closeCurrentOnError();
 }