+++ src/main/java/com/arangodb/model/AqlQueryOptions.java
@@ -26,6 +26,7 @@ import java.util.Arrays;
 import java.util.Collection;
 
 import com.arangodb.velocypack.VPackSlice;
+import com.arangodb.velocypack.annotations.Expose;
 
 /**
  * @author Mark Vollmary
@@ -45,6 +46,8 @@ public class AqlQueryOptions implements Serializable {
 	private VPackSlice bindVars;
 	private String query;
 	private Options options;
+	@Expose(serialize = false)
+	private Boolean allowDirtyRead;
 
 	public AqlQueryOptions() {
 		super();
@@ -433,6 +436,23 @@ public class AqlQueryOptions implements Serializable {
 
 	private static class Optimizer {
 		private Collection<String> rules;
+	}
+
+	/**
+	 * @see <a href="https://docs.arangodb.com/current/Manual/Administration/ActiveFailover/#reading-from-follower">API
+	 *      Documentation</a>
+	 * @param allowDirtyRead
+	 *            Set to {@code true} allows reading from followers in an active-failover setup.
+	 * @since ArangoDB 3.4.0
+	 * @return options
+	 */
+	public AqlQueryOptions allowDirtyRead(final Boolean allowDirtyRead) {
+		this.allowDirtyRead = allowDirtyRead;
+		return this;
+	}
+
+	public Boolean getAllowDirtyRead() {
+		return allowDirtyRead;
 	}
 
 }